+++ reconstruction/src/instance/facts.rs
@@ -13,17 +13,17 @@ use super::{
 
 #[derive(Debug, PartialEq)]
 struct PredicateFacts {
-    index_map: HashMap<Vec<usize>, usize>,
+    index_map: HashMap<Vec<u32>, u32>,
 }
 impl PredicateFacts {
     fn new(
         types: &Option<Types>,
         objects: &Objects,
-        predicate: usize,
-        statics: &Vec<(usize, Vec<usize>)>,
+        predicate: u32,
+        statics: &Vec<(u32, Vec<u32>)>,
         is_static: bool,
         parameters: &Parameters,
-        offset: usize,
+        offset: u32,
     ) -> Self {
         let permutations = match is_static {
             true => statics
@@ -35,9 +35,9 @@ impl PredicateFacts {
                 .collect(),
             false => permute_mutable(types, objects, &parameters.parameter_types),
         };
-        let mut index_map: HashMap<Vec<usize>, usize> = HashMap::new();
+        let mut index_map: HashMap<Vec<u32>, u32> = HashMap::new();
         for permutation in permutations.into_iter() {
-            index_map.insert(permutation, offset + index_map.len());
+            index_map.insert(permutation, offset + index_map.len() as u32);
         }
         Self { index_map }
     }
@@ -46,11 +46,11 @@ impl PredicateFacts {
         self.index_map.len()
     }
 
-    fn index(&self, permutation: &Vec<usize>) -> usize {
+    fn index(&self, permutation: &Vec<u32>) -> u32 {
         self.index_map[permutation]
     }
 
-    fn get_permutation(&self, index: usize) -> &Vec<usize> {
+    fn get_permutation(&self, index: u32) -> &Vec<u32> {
         self.index_map
             .iter()
             .find(|(_, i)| **i == index)
@@ -58,7 +58,7 @@ impl PredicateFacts {
             .unwrap()
     }
 
-    fn contains(&self, permutation: &Vec<usize>) -> bool {
+    fn contains(&self, permutation: &Vec<u32>) -> bool {
         self.index_map.contains_key(permutation)
     }
 }
@@ -66,8 +66,8 @@ impl PredicateFacts {
 #[derive(Debug, PartialEq)]
 pub struct Facts {
     facts: Vec<PredicateFacts>,
-    static_predicates: HashSet<usize>,
-    init: HashSet<usize>,
+    static_predicates: HashSet<u32>,
+    init: HashSet<u32>,
 }
 
 impl Facts {
@@ -79,35 +79,37 @@ impl Facts {
         inits: &Inits,
     ) -> Self {
         let mut facts: Vec<PredicateFacts> = Vec::new();
-        let mut static_predicates: HashSet<usize> = HashSet::new();
+        let mut static_predicates: HashSet<u32> = HashSet::new();
 
-        let mut statics: Vec<(usize, Vec<usize>)> = Vec::new();
+        let mut statics: Vec<(u32, Vec<u32>)> = Vec::new();
         for init in inits.iter() {
             let predicate = World::global().get_predicate_index(&init.name);
             let parameters = World::global().get_object_indexes(&init.parameters);
             statics.push((predicate, parameters));
         }
 
-        let mut offset: usize = 0;
+        let mut offset: u32 = 0;
         for (i, predicate) in predicates.predicate_parameters().iter().enumerate() {
-            let predicate_name = World::global().get_predicate_name(i);
+            let predicate_name = World::global().get_predicate_name(i as u32);
             status_print(Status::Init, &format!("Grounding: {}", predicate_name));
             // TODO: Make sure this works!
-            let is_static = check_static_all(actions, i) || check_degrading_all(actions, i);
+            let is_static =
+                check_static_all(actions, i as u32) || check_degrading_all(actions, i as u32);
             println!("is static: {}", is_static);
             if is_static {
-                static_predicates.insert(i);
+                static_predicates.insert(i as u32);
             }
-            let predicate_facts =
-                PredicateFacts::new(types, objects, i, &statics, is_static, predicate, offset);
+            let predicate_facts = PredicateFacts::new(
+                types, objects, i as u32, &statics, is_static, predicate, offset,
+            );
             println!("facts: {}", predicate_facts.count());
-            offset += predicate_facts.count();
+            offset += predicate_facts.count() as u32;
             facts.push(predicate_facts);
         }
 
-        let init: HashSet<usize> = statics
+        let init: HashSet<u32> = statics
             .iter()
-            .map(|(p, par)| facts[*p].index(par))
+            .map(|(p, par)| facts[*p as usize].index(par))
             .collect();
         Self {
             facts,
@@ -116,36 +118,36 @@ impl Facts {
         }
     }
 
-    pub fn index(&self, predicate: usize, parameters: &Vec<usize>) -> usize {
-        self.facts[predicate].index(parameters)
+    pub fn index(&self, predicate: u32, parameters: &Vec<u32>) -> u32 {
+        self.facts[predicate as usize].index(parameters)
     }
 
-    pub fn fact_predicate(&self, fact_index: usize) -> usize {
+    pub fn fact_predicate(&self, fact_index: u32) -> u32 {
         let mut acc = 0;
-        for i in 0..self.facts.len() {
-            acc += self.facts[i].count();
+        for i in 0..self.facts.len() as u32 {
+            acc += self.facts[i as usize].count() as u32;
             if fact_index < acc {
                 return i;
             }
         }
-        self.facts.len() - 1
+        self.facts.len() as u32 - 1
     }
 
-    pub fn fact_parameters(&self, fact_index: usize) -> &Vec<usize> {
+    pub fn fact_parameters(&self, fact_index: u32) -> &Vec<u32> {
         let predicate = self.fact_predicate(fact_index);
-        let facts = &self.facts[predicate];
+        let facts = &self.facts[predicate as usize];
         facts.get_permutation(fact_index)
     }
 
-    pub fn is_static(&self, predicate: usize) -> bool {
+    pub fn is_static(&self, predicate: u32) -> bool {
         self.static_predicates.contains(&predicate)
     }
 
-    pub fn is_statically_true(&self, predicate: usize, parameters: &Vec<usize>) -> bool {
-        self.facts[predicate].contains(parameters)
+    pub fn is_statically_true(&self, predicate: u32, parameters: &Vec<u32>) -> bool {
+        self.facts[predicate as usize].contains(parameters)
     }
 
-    pub fn get_static_true(&self) -> Vec<usize> {
+    pub fn get_static_true(&self) -> Vec<u32> {
         self.init
             .iter()
             .filter_map(|i| match !self.is_static(self.fact_predicate(*i)) {
@@ -155,7 +157,7 @@ impl Facts {
             .collect()
     }
 
-    pub fn get_init(&self) -> HashSet<usize> {
+    pub fn get_init(&self) -> HashSet<u32> {
         self.init
             .iter()
             .filter_map(|i| match self.is_static(self.fact_predicate(*i)) {
@@ -166,7 +168,7 @@ impl Facts {
     }
 }
 
-fn check_static(predicate: usize, exp: &Expression) -> bool {
+fn check_static(predicate: u32, exp: &Expression) -> bool {
     match exp {
         Expression::Predicate { index, .. } => *index != predicate,
         Expression::And(exp) => exp.iter().all(|exp| check_static(predicate, exp)),
@@ -175,14 +177,14 @@ fn check_static(predicate: usize, exp: &Expression) -> bool {
     }
 }
 
-fn check_static_all(actions: &Actions, predicate: usize) -> bool {
+fn check_static_all(actions: &Actions, predicate: u32) -> bool {
     actions
         .actions
         .iter()
         .all(|a| check_static(predicate, &a.effect))
 }
 
-fn check_degrading(predicate: usize, exp: &Expression, val: bool) -> bool {
+fn check_degrading(predicate: u32, exp: &Expression, val: bool) -> bool {
     match exp {
         Expression::Predicate { index, .. } => *index != predicate || !val,
         Expression::And(exp) => exp.iter().all(|exp| check_degrading(predicate, exp, val)),
@@ -191,7 +193,7 @@ fn check_degrading(predicate: usize, exp: &Expression, val: bool) -> bool {
     }
 }
 
-fn check_degrading_all(actions: &Actions, predicate: usize) -> bool {
+fn check_degrading_all(actions: &Actions, predicate: u32) -> bool {
     actions
         .actions
         .iter()
