+++ reconstruction/src/instance/operator.rs
@@ -8,14 +8,14 @@ use super::{actions::Action, expression::Expression, permute::permute_mutable, I
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Operator {
-    pub pre_pos: HashSet<usize>,
-    pub pre_neg: HashSet<usize>,
-    pub eff_pos: HashSet<usize>,
-    pub eff_neg: HashSet<usize>,
+    pub pre_pos: HashSet<u32>,
+    pub pre_neg: HashSet<u32>,
+    pub eff_pos: HashSet<u32>,
+    pub eff_neg: HashSet<u32>,
 }
 
 impl Operator {
-    pub fn get_effect(&self) -> Vec<(usize, bool)> {
+    pub fn get_effect(&self) -> Vec<(u32, bool)> {
         let mut effect = vec![];
 
         for i in self.eff_pos.iter() {
@@ -32,13 +32,13 @@ impl Operator {
 
 fn extract_from_action(
     instance: &Instance,
-    parameters: &Vec<usize>,
+    parameters: &Vec<u32>,
     action: &Action,
 ) -> Option<Operator> {
-    let mut pre_pos: HashSet<usize> = HashSet::new();
-    let mut pre_neg: HashSet<usize> = HashSet::new();
-    let mut eff_pos: HashSet<usize> = HashSet::new();
-    let mut eff_neg: HashSet<usize> = HashSet::new();
+    let mut pre_pos: HashSet<u32> = HashSet::new();
+    let mut pre_neg: HashSet<u32> = HashSet::new();
+    let mut eff_pos: HashSet<u32> = HashSet::new();
+    let mut eff_neg: HashSet<u32> = HashSet::new();
     if let Some(exp) = &action.precondition {
         if !walk(instance, parameters, &mut pre_pos, &mut pre_neg, exp, true) {
             return None;
@@ -68,14 +68,14 @@ pub fn generate_operator_string(
     parameters: &Vec<String>,
 ) -> Operator {
     let action: &Action = instance.get_action(action);
-    let parameters: Vec<usize> = World::global().get_object_indexes(parameters);
+    let parameters: Vec<u32> = World::global().get_object_indexes(parameters);
     extract_from_action(instance, &parameters, action).unwrap()
 }
 
 pub fn generate_operators<'a>(
     instance: &'a Instance,
     action: &'a Action,
-) -> impl Iterator<Item = (Operator, Vec<usize>)> + 'a {
+) -> impl Iterator<Item = (Operator, Vec<u32>)> + 'a {
     let permutations = permute_mutable(
         &instance.types,
         &instance.objects,
@@ -89,9 +89,9 @@ pub fn generate_operators<'a>(
 
 fn walk(
     instance: &Instance,
-    permutation: &Vec<usize>,
-    pos: &mut HashSet<usize>,
-    neg: &mut HashSet<usize>,
+    permutation: &Vec<u32>,
+    pos: &mut HashSet<u32>,
+    neg: &mut HashSet<u32>,
     exp: &Expression,
     value: bool,
 ) -> bool {
@@ -99,7 +99,10 @@ fn walk(
         Expression::Predicate { index, parameters } => {
             let exp = if value { pos } else { neg };
 
-            let parameters = parameters.iter().map(|p| permutation[*p]).collect_vec();
+            let parameters = parameters
+                .iter()
+                .map(|p| permutation[*p as usize])
+                .collect_vec();
 
             match instance.facts.is_static(*index) {
                 true => {
